Change the way starter-kit-org adds the org info manual
Changed how the starter-kit adds the info manual to the info tree. Jambunathan K mentioned on the emacs-orgmode@gnu.org mailing list that the official org-mode ELPA tarball adds the new version of the info manual to Info-directory-list using the "push" function.

This is much more direct and cleaner (IMO) way of doing things. I think the info on the "push" function justifies the change:

,----[ C-h f push RET ]
| push is a Lisp macro in `cl.el'.
|
| (push X PLACE)
|
| Insert X at the head of the list stored in PLACE.
| Analogous to (setf PLACE (cons X PLACE)), though more careful about
| evaluating each argument only once and in the right order.  PLACE may
| be a symbol, or any generalized variable allowed by `setf'.
`----

It is not clear to me that copying Info-default-directory-list to Info-directory-list is required. Users should not be changing Info-default-directory-list in general, but it is nonetheless copied to Info-directory-list *iff* the latter variable is empty (has been that way for a few emacs versions). "push" is a non-destructive way to put a new directory to the top of Info-directory-list.